Show message when no related programs were found
Other changes:

- Lint code
- Put all GQL queries closer to their clients
