+++ src/upgrade/client.ts
@@ -1,6 +1,19 @@
 import Client from "../client";
-import { queryLatestCliVersion } from "../graphql";
 import { LatestCliReleaseResponse } from "../interfaces";
+
+const queryLatestCliVersion = `
+query CliVersions($cliVersion: String) {
+  latestCliRelease(cliVersion: $cliVersion) {
+    isCliVersionCurrent
+    latestRelease {
+      url
+			tagName
+      publishedAt
+      body
+    }
+  }
+}
+`;
 
 export default class UpgradeClient extends Client {
   constructor() {
